Worked a bit more on character movement. Also fixed a bunch of warnings.

+++ SFML Test/SFML Test/model/bullet.cpp
@@ -53,7 +53,7 @@ namespace mp
 
 		// Calculate unique identifying bits for this bullet.
 		// (This makes us pass through our player)
-		const short playerBits = pow(2.0, owner + 1);
+		const short playerBits = short(pow(2.0, owner + 1));
 		fixtureDef.filter.categoryBits = playerBits;
 		fixtureDef.filter.maskBits = 0xFFFF & (~playerBits);
